# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de ejemplo",
        "description": "Documentación de la API de ejemplo",
        "contact": {
            "email": "contacto@example.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/accountable/access": {
            "post": {
                "tags": [
                    "Responsable"
                ],
                "summary": "Access to accountable data with JWT token",
                "description": "Authenticate a accountable by CI and code, then generate a JWT token.",
                "operationId": "access",
                "requestBody": {
                    "description": "Credentials to access accountable data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ci",
                                    "code"
                                ],
                                "properties": {
                                    "ci": {
                                        "type": "integer",
                                        "example": 12345678,
                                        "deprecated": true
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "NFvyeVUh fEChWv2q J2zu3xYd RfVYrsGB njvFM92m",
                                        "deprecated": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "generated.jwt.token",
                                            "deprecated": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid code",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid code",
                                            "deprecated": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Responsable not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Responsable not found",
                                            "deprecated": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "deprecated": true,
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "example": "Error message"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}